Fix Issues with Initialization Tab GUI Code
